requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
beautifulsoup4==4.12.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
confection==0.1.0
cryptography==41.0.2
cymem==2.0.7
dataclasses-json==0.5.13
decorator==5.1.1
dill==0.3.7
dnspython==2.4.1
elastic-transport==8.4.0
elasticsearch==8.9.0
exceptiongroup==1.1.2
faiss-cpu==1.7.4
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gensim==4.3.1
google-api-core==2.11.1
google-api-python-client==2.70.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
grpcio-tools==1.56.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
jaraco.context==4.3.0
Jinja2==3.1.2
joblib==1.3.1
langchain==0.0.247
langcodes==3.3.0
langsmith==0.0.15
loguru==0.7.0
manifest-ml==0.0.1
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.0.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pathy==0.10.2
pdfminer.six==20221105
pinecone-client==2.2.2
preshed==3.0.8
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pyparsing==3.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
qdrant-client==0.11.10
redis==4.6.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
safetensors==0.3.1
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.49
sqlitedict==2.1.0
srsly==2.4.7
tenacity==8.2.2
thinc==8.1.10
tiktoken==0.4.0
tokenizers==0.13.3
torch==1.13.1
tqdm==4.65.0
transformers==4.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
validators==0.20.0
wasabi==1.1.2
weaviate-client==3.22.1
wikipedia==1.4.0
win32-setctime==1.1.0
wolframalpha==5.0.0
xmltodict==0.13.0
yarl==1.9.2
